Fix type errors in tests

+++ tests/listComprehensions.hs
@@ -1,4 +1,4 @@
-import Prelude
+import           Prelude
 
 main :: Fay ()
-main = putStrLn $ show $ sum [ x*x | x <- [1, 2, 3, 4, 5], let y = x + 4, y < 8]
+main = putStrLn $ show $ sum [ x*x | x <- [1::Int, 2, 3, 4, 5], let y = x + 4, y < 8]
